Feat: 자동차 경주 게임을 완료한 후 누가 우승했는지를 알려준다.
